Tweaked Models, Made create_user.py
As we aren't going to be allowing public signups, have made a helper
script for creating users. It has no error handling, so is rough and
ready, but the peewee error messages should suffice.
It sends a message with the register link in.

Models - password can now be null, but this *shouldn't* matter as it
will only be able to be set in either the register or change views, and
they will be validated.